requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
playwright==1.58.0
